Extract the whisper server into a separate class
